Refactor user deletion endpoint to improve authorization checks and simplify route structure

+++ src/controllers/user.controller.js
@@ -4,14 +4,16 @@ import { User } from "../models/user.model.js"
 
 export const deleteUserController = async (req, res) => {
 	const userId = req.userData.userId
-
 	const userData = await User.findById(userId)
 
-	if (!userData) return res.status(404).json({ message: `User with ID ${userId} does not exists.` });
-
-	await userData.deleteOne({ _id: userId })
-
-	res.json({ message: `User with ID ${userId} deleted successfully.` });
+	try {
+		if (!userData) return res.status(404).json({ message: `User with ID ${userId} does not exists.` });
+		if (userData.whatRole !== 'masterAdmin') return res.status(401).json({ success: false, message: 'You are not authorized to perform this action. This require Admin Authorization.' });
+		await userData.deleteOne({ _id: userId })
+		res.json({ message: `User with ID ${userId} deleted successfully.` });
+	} catch (error) {
+		res.status(500).json({ message: error.message });
+	}
 }
 
 export const newRefreshTokens = async (req, res) => {
